Merges origin/main (37 commits) into our fork: upstream removed the
MpvMediaPlayer wrapper and rewrote the native event model.

Resolution policy:
- native layer (MpvPlayer.idl/.h/.cpp, MpvPreviewer.*, vcxproj, packages):
  taken from upstream (new event model + InitializeForPreview + PlaylistXxx
  + WindowsAppSDK 2.3.9/2.3.6); kept our ApiAndFunc additions where they
  do not overlap
- MpvMediaPlayer*.cs (4 files): accepted upstream deletion
- settings option controls + README: taken from upstream (their UI refactor)
- AppContext/csproj/AppSettings/AppLang/window layer/settings pages: ours
  kept (fork features + newly added files such as ControlBarAdaptiveLayout)
- player pages conflicting with our fork implementations (menu, preview,
  disc open, playlist, page XAML): taken from upstream per the decision to
  follow upstream's architecture; our fork features in those files need a
  follow-up pass

Known follow-ups (next phase, see the worklist doc):
- C# adaptation: '_mediaPlayer.Native.X' -> '_mediaPlayer.X' (12 sites),
  '_mediaPlayer.Close()' -> 'Destroy()', 'MediaOpened' -> 'FileLoaded'
- re-insert this session's fixes: aspect fit + panel size publish,
  display-peak HDR gate, control bar tiers + overflow, throttle
- build and fix the fallout, then regression before touching main

main is untouched and remains the working build.

+++ mpv-winui/mpv-winui/Modules/Player/MpvPlayerPage_Display.cs
@@ -496,7 +496,7 @@ namespace mpv_winui.Modules.Player
                     _lastRefreshRate = rate;
                     if (ShouldAutoApplyRefreshRate())
                     {
-                        _mediaPlayer?.UpdateDisplayRefreshRate(rate);
+                        _mediaPlayer?.UpdateDisplayRefreshRate((int)rate);
                     }
                 }
                 ApplyDisplayPeak();
